+++ src/main/kotlin/com/naltynbekkz/oneaviation/plane/CreatePlaneRequest.kt
@@ -2,7 +2,7 @@ package com.naltynbekkz.oneaviation.plane
 
 data class CreatePlaneRequest(
     val name: String,
-    val mileage: Long,
+    val mileage: Int,
     val capacity: Int,
     val price: Float,
 )
