+++ src/main/kotlin/com/naltynbekkz/oneaviation/user/UserRepository.kt
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UserRepository : JpaRepository<UserEntity, Long> {
+interface UserRepository : JpaRepository<UserEntity, Int> {
     fun findByUsername(email: String): UserEntity?
 
     @Query(
